Refactor: update of package name from 'Interfaces' to 'ports'
