+++ src/main/kotlin/HttpAPI.kt
@@ -3,14 +3,10 @@ import org.http4k.core.Method.POST
 import org.http4k.core.Method.PUT
 import org.http4k.core.Method.DELETE
 import org.http4k.core.Status.Companion.OK
-import org.http4k.filter.DebuggingFilters.PrintRequest
 import org.http4k.routing.bind
 import org.http4k.routing.routes
-import org.http4k.server.SunHttp
-import org.http4k.server.asServer
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-import Interfaces.TodoListRepo
 import org.http4k.core.*
 import org.http4k.routing.path
